Code cleanup

+++ scenes/components/LootTableItem.cs
@@ -2,7 +2,7 @@ using Godot;
 
 public partial class LootTableItem : Node
 {
-	[Export] public int Weight { get; private set; } = 1;
+	[Export] public float Weight { get; private set; } = 1;
 
 	[Export] public PackedScene Scene { get; private set; }
 }
